Fixed integer value in Revit Version 2024

+++ Common_glTF_Exporter/Core/GLTFExtras.cs
@@ -15,7 +15,7 @@
 
         public Dictionary<string, string> parameters { get; set; }
 
-        public int elementId { get; set; }
+        public long elementId { get; set; }
 
         public string elementCategory { get; set; }
     }
